+++ ThemesAndStyles.c
@@ -1318,8 +1318,8 @@ void theme_prepareChildrenForDarkmode(HWND hParent) {
  * Returns the foreground color of a text style with a given name in the current theme.
  * If the style cannot be found a default value is returned.
  */
-long theme_textStyleForeground(const char* pszStylename, long lDefault) {
-	EDTEXTSTYLE* pStyle = ll_find((void*)theme_getCurrent()->th_styles, pszStylename);
+long theme_textStyleForeground(THEME_DATA* pTheme, const char* pszStylename, long lDefault) {
+	EDTEXTSTYLE* pStyle = ll_find((void*)pTheme->th_styles, pszStylename);
 	return pStyle ? pStyle->fgcolor : lDefault;
 }
 
@@ -1327,8 +1327,8 @@ long theme_textStyleForeground(const char* pszStylename, long lDefault) {
  * Returns the background color of a text style with a given name in the current theme.
  * If the style cannot be found a default value is returned.
  */
-long theme_textStyleBackground(const char* pszStylename, long lDefault) {
-	EDTEXTSTYLE* pStyle = ll_find((void*)theme_getCurrent()->th_styles, pszStylename);
+long theme_textStyleBackground(THEME_DATA* pTheme, const char* pszStylename, long lDefault) {
+	EDTEXTSTYLE* pStyle = ll_find((void*)pTheme->th_styles, pszStylename);
 	return pStyle ? pStyle->bgcolor : lDefault;
 }