+++ include/themes.h
@@ -24,7 +24,7 @@ typedef struct tagTHEME_DATA {
 	int		 th_isWinTheme;					// true for the default (Windows Theme)
 	int		 th_isDarkMode;					// true for all themes, that should use darkmode for window NC areas etc...
 	COLORREF th_iconColor;					// color of the icons in the toolbar.
-	COLORREF th_defaultBackgroundColor;
+	COLORREF th_defaultBackgroundColor;		// standard background color of text editor text.
 	COLORREF th_changedLineColor;
 	COLORREF th_savedChangedLineColor;
 	COLORREF th_undoAfterSaveChangedLineColor;
@@ -125,13 +125,13 @@ extern void font_setDefaultTextColors(HDC hdc, THEME_DATA* pTheme);
  * Returns the foreground color of a text style with a given name in the current theme.
  * If the style cannot be found a default value is returned.
  */
-extern long theme_textStyleForeground(const char* pszStylename, long lDefault);
+extern long theme_textStyleForeground(THEME_DATA* pTheme, const char* pszStylename, long lDefault);
 
 /*
  * Returns the background color of a text style with a given name in the current theme.
  * If the style cannot be found a default value is returned.
  */
-long theme_textStyleBackground(const char* pszStylename, long lDefault);
+long theme_textStyleBackground(THEME_DATA* pTheme, const char* pszStylename, long lDefault);
 
 /*
  * Returns the font face for a given style name. If the style cannot be found or does not